+++ api/src/certification/enrolment/routes.js
@@ -2,6 +2,7 @@ import * as attendanceSheet from './application/attendance-sheet-route.js';
 import * as certificationCandidate from './application/certification-candidate-route.js';
 import * as companion from './application/companion-route.js';
 import * as country from './application/country-route.js';
+import * as enrolment from './application/enrolment-route.js';
 import * as sessionMassImport from './application/session-mass-import-route.js';
 import * as session from './application/session-route.js';
 
@@ -10,6 +11,7 @@ const certificationEnrolmentRoutes = [
   certificationCandidate,
   companion,
   country,
+  enrolment,
   session,
   sessionMassImport,
 ];
